+++ lib/glimmer/libui/control_proxy/area_proxy.rb
@@ -34,9 +34,13 @@ module Glimmer
         class << self
           # this attribute is only populated during on_draw call
           attr_accessor :current_area_draw_params
+          
+          def ext_key_symbols
+            @ext_key_symbols ||= ::LibUI.constants.select { |c| c.to_s.start_with?('ExtKey') }.map { |c| c.to_s.underscore.sub('ext_key_', '') }
+          end
         end
         
-        LISTENERS = ['on_draw', 'on_mouse_event', 'on_mouse_move', 'on_mouse_down', 'on_mouse_up', 'on_mouse_drag_start', 'on_mouse_drag', 'on_mouse_drop', 'on_mouse_crossed', 'on_mouse_enter', 'on_mouse_exit', 'on_drag_broken']
+        LISTENERS = ['on_draw', 'on_mouse_event', 'on_mouse_move', 'on_mouse_down', 'on_mouse_up', 'on_mouse_drag_start', 'on_mouse_drag', 'on_mouse_drop', 'on_mouse_crossed', 'on_mouse_enter', 'on_mouse_exit', 'on_drag_broken', 'on_key_event', 'on_key_down', 'on_key_up']
         LISTENER_ALIASES = {
           on_drawn: 'on_draw',
           on_mouse_moved: 'on_mouse_move',
@@ -110,7 +114,16 @@ module Glimmer
           @area_handler.DragBroken   = fiddle_closure_block_caller(0, [1, 1]) do |_, _|
             on_drag_broken.each { |listener| listener.call }
           end
-          @area_handler.KeyEvent     = fiddle_closure_block_caller(0, [0]) {}
+          @area_handler.KeyEvent     = fiddle_closure_block_caller(0, [1, 1, 1]) do |_, _, area_key_event|
+            area_key_event = ::LibUI::FFI::AreaKeyEvent.new(area_key_event)
+            area_key_event = area_key_event_hash(area_key_event)
+            on_key_event.each { |listener| listener.call(area_key_event) }
+            if area_key_event[:up]
+              on_key_up.each { |listener| listener.call(area_key_event) }
+            else
+              on_key_down.each { |listener| listener.call(area_key_event) }
+            end
+          end
         end
         
         def area_draw_params_hash(area_draw_params)
@@ -139,6 +152,24 @@ module Glimmer
           }
         end
         
+        def area_key_event_hash(area_key_event)
+          {
+            key: key_to_char(area_key_event.Key),
+            ext_key: ext_key_to_symbol(area_key_event.ExtKey),
+            modifier: modifiers_to_symbols(area_key_event.Modifier).first,
+            modifiers: modifiers_to_symbols(area_key_event.Modifiers),
+            up: Glimmer::LibUI.integer_to_boolean(area_key_event.Up),
+          }
+        end
+        
+        def key_to_char(key)
+          key.chr if key > 0
+        end
+        
+        def ext_key_to_symbol(ext_key_value)
+          AreaProxy.ext_key_symbols[ext_key_value - 1].to_s.to_sym if ext_key_value > 0
+        end
+                
         def modifiers_to_symbols(modifiers_value)
           symbols = []
           modifiers_value = extract_symbol_from_modifiers_value(modifiers_value, symbols: symbols) while modifiers_value > 0